requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.1.0  # https://github.com/evansd/whitenoise
unidecode==1.1.1 # https://pypi.org/project/Unidecode/

# Django
# ------------------------------------------------------------------------------
django==3.0.8  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-autoslug==1.9.7  # https://github.com/justinmayer/django-autoslug/
djangorestframework==3.11.0
